relations: various updates
Add two char relations <> >= <= and also ** for power
Add string version of relations

Not implemented power yet - just the parsing for it

Add notes on errors in the relation handling. We need another layer of
logic to correctly parse A = B AND C = D, plus we probably want to split
the current funky & | C style ops from AND / OR as boolean ops instead.

(We may also want to dump the current uses - lots of basics use & for something
 else)